Added code in `supermom.test.c` to check the success count.
This is written in a way that should work no matter the supermom configuration.
When running as root and the supermom module creates `superlog/success`, it successfully reads and checks the success count.
When not running as root, it can't read `superlog/success`, but none of the `supermom` calls will succeed anyways.
When the supermom module doesn't create `superlog/success`, like in part4, it can't read it because it doesn't exist.
Either way, the corresponding correct behavior is then checked.